+++ old_test/pack_game/_utils/game_level_generator_test.dart
@@ -1,8 +1,6 @@
-import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:word_by_word_game/_abstract/abstract.dart';
-import 'package:word_by_word_game/pack_game/pack_game.dart';
 
 void main() {
   group('GameLevelGenerator', () {
@@ -14,16 +12,16 @@ void main() {
       final gamePlayer = player.toNewGameProfile();
       final levelGenerator = GameLevelGenerator();
       final shelfLevels = levelGenerator.createBookShelfLevels(
-        players: [gamePlayer],
+        characters: [gamePlayer],
         screenWidth: 600,
       );
 
       expect(shelfLevels.isNotEmpty, isTrue);
 
       final shelfLevel = shelfLevels.first;
-      expect(shelfLevel.players.isNotEmpty, isTrue);
+      expect(shelfLevel.characters.isNotEmpty, isTrue);
       expect(
-        shelfLevel.players
+        shelfLevel.characters
             .firstWhereOrNull((final player) => player == gamePlayer),
         isNotNull,
       );