+++ packages/wbw_core/lib/src/models/player_character_model.dart
@@ -1,0 +1,25 @@
+// ignore_for_file: invalid_annotation_target
+
+part of 'models.dart';
+
+typedef PlayerCharacterModelId = String;
+
+@immutable
+@Freezed(
+  fromJson: true,
+  toJson: true,
+  equal: true,
+  addImplicitFinal: true,
+  copyWith: true,
+)
+class PlayerCharacterModel with _$PlayerCharacterModel {
+  @JsonSerializable(explicitToJson: true)
+  const factory PlayerCharacterModel({
+    required final PlayerCharacterModelId id,
+    required final String name,
+    required final String description,
+  }) = _PlayerCharacterModel;
+  const PlayerCharacterModel._();
+  factory PlayerCharacterModel.fromJson(final Map<String, dynamic> json) =>
+      _$PlayerCharacterModelFromJson(json);
+}